refactor: divider types
