+++ projects/ion/src/lib/core/types/divider.ts
@@ -1,0 +1,10 @@
+import { DirectionType } from './direction';
+
+export type DividerType = 'solid' | 'dashed' | 'text';
+
+export interface IonDividerProps {
+  label?: string;
+  direction?: DirectionType;
+  type?: DividerType;
+  margin?: boolean;
+}
