+++ projects/ion/src/lib/divider/divider.component.ts
@@ -1,15 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { DirectionType } from '../core/types';
-
-type DividerType = 'solid' | 'dashed' | 'text';
-
-export interface IonDividerProps {
-  label?: string;
-  direction?: DirectionType;
-  type?: DividerType;
-  margin?: boolean;
-}
-
+import { DividerType } from '../core/types/divider';
 @Component({
   selector: 'ion-divider',
   templateUrl: './divider.component.html',
